+++ docs/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 ### Changed
+- Improve responsiveness on startup.
+
 ### Fixed
 
 ## [3.21.0] - 2023-12-04